+++ R/knitr-hooks.R
@@ -3,9 +3,8 @@
 #' Refer to [knitr documentation](https://yihui.org/knitr/hooks).  This gives us
 #' an opportunity to rewrite a chunk's output at the knit.md stage.  In particular,
 #' we use this for images:
-#' 
-#'   HTML images are encoded as base64 and wrapped in an img tag;
-#'   PDF images are referenced with a markdown construction, e.g. "![]()"
+#' * HTML images are encoded as base64 and wrapped in an img tag
+#' * PDF images are referenced with a markdown construction
 #' 
 #' @keywords internal
 #' @param before a boolean