+++ R/utils-wrap.R
@@ -1,8 +1,5 @@
-#!/usr/bin/env -S Rscript --vanilla  # nolint
 
-library(magrittr)
-
-ops <- list(
+cliops <- list(
   info = list(),
   warn = list(
     "cli_alert_warning",
@@ -27,9 +24,12 @@ ops <- list(
     "cli_progress_builtin_handlers",
     "cli_progress_cleanup",
     "cli_progress_demo",
-    "cli_progress_done",    
+    "cli_progress_done",   
+    "cli_progress_message",
     "cli_progress_num",
+    "cli_progress_output", 
     "cli_progress_styles",
+    "cli_progress_step",
     "cli_progress_update",
     "cli_sitrep",
     "cli_status",
@@ -39,13 +39,8 @@ ops <- list(
   )
 )
 
-cli_exports <- getNamespaceExports("cli")
-cli_names <- grep("^cli_", cli_exports, value = TRUE) %>%  
-  setdiff(ops$omit) %>%
-  sort() 
-
 get_level <- function(ops, name) {
-  # check name against each list in ops; first category (info) is the defacto default
+  # check name against each list in cliops; first category (info) is the defacto default
   key <- names(which.max(sapply(ops, function(l) name %in% l)))
   switch(
     key,
@@ -55,22 +50,33 @@ get_level <- function(ops, name) {
   )
 }
 
-cli_levels <- cli_names %>%
-  rlang::set_names() %>%
-  purrr::imap(
-    \(n, i) list(name = i, level = get_level(ops, n))
-  )
-names(cli_levels) <- NULL
+#' build the AUTO-cli-wrappers.R file
+#' 
+#' @keywords internal
+gen_cli_wrappers <- function() {
 
-# render the template
-whisker::whisker.render(
-  xfun::read_utf8(
-    here::here("./templates/cli_wrappers.tmpl")
-  ), 
-  data = list(
-    ops = cli_levels
-  )
-) %>%
-  xfun::write_utf8(
-    here::here("./R/AUTO-cli-wrappers.R")    
-  )
+  cli_exports <- getNamespaceExports("cli")
+  cli_names <- grep("^cli_", cli_exports, value = TRUE) %>%  
+    setdiff(cliops$omit) %>%
+    sort() 
+
+  cli_levels <- cli_names %>%
+    rlang::set_names() %>%
+    purrr::imap(
+      \(n, i) list(name = i, level = get_level(cliops, n))
+    )
+  names(cli_levels) <- NULL
+
+  # render the template
+  whisker::whisker.render(
+    xfun::read_utf8(
+      here::here("./templates/cli_wrappers.tmpl")
+    ), 
+    data = list(
+      ops = cli_levels
+    )
+  ) %>%
+    xfun::write_utf8(
+      here::here("./R/AUTO-cli-wrappers.R")    
+    )
+}